+++ src/main/java/com/react/pnld/services/FileService.java
@@ -1,7 +1,7 @@
 package com.react.pnld.services;
 
-import com.react.pnld.model.ScheduleFileLoadDTO;
-import com.react.pnld.model.ProcesaArchivoDTO;
+import com.react.pnld.model.dto.ScheduleFileLoadDTO;
+import com.react.pnld.model.dto.ProcesaArchivoDTO;
 import com.react.pnld.model.ScheduleFileLoadResponse;
 import com.react.pnld.repo.FileRepository;
 import org.slf4j.Logger;
